+++ python/ml_framework/handover_training.py
@@ -13,23 +13,16 @@ import plotly.graph_objects as go
 import math
 import time
 import scipy.signal
-import git
 
-from python.ml_framework.data.MotionPreprocessing import MotionProcessor, count_lines, read_csv_data, ReadBinary 
+from python.ml_framework.data.motion_preprocessing import MotionProcessor, count_lines, read_csv_data, ReadBinary 
 
-def get_git_root():
-    """Returns the root directory of the git repository containing the current working directory."""
-    try:
-        repo = git.Repo(search_parent_directories=True)
-        return repo.working_tree_dir
-    except git.exc.InvalidGitRepositoryError:
-        return None
     
 def main():
     # common config
-    dataset_path = get_git_root() + r"\datasets\Survivor_Gen"
+    # dataset_path = r"C:\DEV\DATASETS\Survivor_Gen"
+    dataset_path = r"C:\anim-ws\AnimHost\datasets\Survivor_Gen"
     # path_to_ai4anim = r"C:\anim-ws\Handover_AnimHost\Handover AnimHost Implementation\AI4Animation_Siggraph2022_Starke\PyTorch"
-    path_to_ai4anim = get_git_root() + r"\AI4Animation\AI4Animation\SIGGRAPH_2022\PyTorch"
+    path_to_ai4anim = r"C:\anim-ws\AI4Animation\AI4Animation\SIGGRAPH_2022\PyTorch"
     mp = MotionProcessor(dataset_path, path_to_ai4anim)   
     # count number of velcocity samples
     num_samples_total = count_lines(dataset_path + "/sequences_velocity.txt")
